requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.5.2
apturl==0.5.2
argon2-cffi==21.1.0
asn1crypto==0.24.0
astunparse==1.6.3
attrs==21.2.0
autograd==1.3
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
bleach==4.1.0
blinker==1.4
blis==0.7.9
Brlapi==0.7.0
CacheControl==0.12.6
cachetools==5.3.2
catalogue==2.0.8
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.3
coloredlogs==7.3
command-not-found==0.3
confection==0.0.3
cryptography==2.8
cupshelpers==1.0
cwltool==2.0.20200224214940
cycler==0.11.0
cymem==2.0.7
DAWG-Python==0.7.2
dbus-python==1.2.16
debugpy==1.5.1
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distro==1.4.0
distro-info==0.23+ubuntu1.1
dnspython==1.16.0
docopt==0.6.2
duplicity==0.8.12.0
en-core-web-lg==3.4.1
en-core-web-sm==3.4.1
entrypoints==0.3
exceptiongroup==1.2.2
execnet==1.9.0
fastapi==0.115.8
fasteners==0.14.1
Flask==2.0.2
flatbuffers==23.5.26
fonttools==4.28.5
future==0.18.2
gast==0.4.0
google-auth==2.27.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gpg==1.13.1
grpcio==1.60.1
h11==0.14.0
h5py==3.10.0
html5lib==1.0.1
httplib2==0.14.0
humanfriendly==4.18
idna==2.8
img2pdf==0.3.3
importlib-resources==5.4.0
iniconfig==1.1.1
ipykernel==6.5.1
ipython==7.30.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isodate==0.6.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
jsonpickle==3.0.4
jsonschema==4.2.1
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.13.1
keyring==18.0.1
keyrings.alt==3.4.0
kiwisolver==1.3.2
langcodes==3.3.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==16.0.6
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
monotonic==1.5
msgpack==0.6.2
murmurhash==1.0.9
mypy-extensions==0.4.3
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
netifaces==0.10.4
networkx==2.4
notebook==6.4.6
numpy==1.22.1
oauth==1.0.1
oauthlib==3.1.0
olefile==0.46
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==20.3
pandas==1.4.0
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.2
pathy==0.10.1
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pluggy==1.0.0
preshed==3.0.8
prometheus-client==0.12.0
prompt-toolkit==3.0.23
protobuf==4.25.2
prov==1.5.2
psutil==5.5.1
psycopg2-binary==2.8.5
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.16.2
pycodestyle==2.8.0
pycparser==2.21
pycrypto==2.6.1
pycups==1.9.73
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.10.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
PyNaCl==1.3.0
pyparsing==2.4.6
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.18.0
pyTelegramBotAPI==4.16.1
pytest==6.2.5
pytest-cache==1.0
pytest-pep8==1.0.6
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2021.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.3.0
qtconsole==5.2.1
QtPy==1.11.2
rdflib==4.2.2
rdflib-jsonld==0.4.0
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.9
ru-core-news-sm==3.4.0
ruamel.yaml==0.15.89
schema-salad==5.0.20200220195218
scipy==1.7.3
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Send2Trash==1.8.0
shellescape==3.4.1
simplejson==3.16.0
sip==4.19.21
six==1.14.0
smart-open==6.3.0
sniffio==1.3.1
soupsieve==1.9.5
spacy==3.4.4
spacy-legacy==3.0.11
spacy-loggers==1.0.4
SPARQLWrapper==1.8.5
srsly==2.4.5
starlette==0.44.0
system-service==0.3
systemd-python==234
telebot==0.0.5
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
terminado==0.12.1
testpath==0.5.0
thinc==8.1.6
toml==0.10.2
tornado==6.1
tqdm==4.64.1
traitlets==5.1.1
typer==0.7.0
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
uvicorn==0.33.0
vsearch==1.0
wadllib==1.3.3
wasabi==0.10.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
wrapt==1.16.0
xkit==0.0.0
zipp==3.6.0
zope.interface==4.7.1
